Move language list UI refresh script to end of language list panel.
The script is placed at the end of the panel so that it will run after the DOM
for the panel has been built, so all elements will definitely be available.

The target element for the checkbox refresh is also changed to be the entire
panel, since it was missing the select-all checkboxes when being run on only
the two lists separately.