Ensure that non-typed changes to filter text trigger filter.
Changes such as cut or paste with the mouse or drag-and-drop were not
triggering filter, since they do not involve a key press. The event
'oninput' appears to be designed for exactly our purpose: to respond
to changes in input text value regardless how the value was changed.

Includes a fallback to respond to typing for browsers that do not
support 'oninput'.