WIP: trying to figure out why projectSlug is null sometimes
When I call getEnabledLocales in the language settings helper, which
defers to the same in ProjectHome, the slug seems to be null. There
are calls either side in which it is not. I don't understand why.

The only solution here may be to make the LanguageSettingsHandler
subclass within ProjectHome completely private, and have all the
calls in the template be to public methods in ProjectHome that
just call a method on the language settings instance.