chore(deps): update dependencies and shrinkwrap
Many shrinkwrap additions here because it was not being run with --dev
before so most of the dependencies were not included. There is a need
to review our policy on which dependencies are included in "dependencies"
and which are included in "devDependencies", as well as which form
of shrinkwrap to run.