chore(deps): update eslint and related dependencies
Several dependencies were updated and a few added to be compatible with the
version of eslint that is now being used.
Note that this introduced some new linting rules, but these have been
disabled to keep the current branch focused on just the dependency changes.
The linting rules will be re-enabled in a later commit.