add separate dev and prod config modules for editor
This is to easily switch on the few differences between the development
build and the production build of the editor. Webpack is responsible for
providing the correct config module to the import call.
All configuration differences in the editor application code must be
included here. It is an error to switch on a production/development flag
anywhere else in the code.
Reasoning: differences between dev and prod builds introduce potential
for production bugs that are not noticed during development. We should
make it easy to see all the differences so that we can accurately assess
this risk. The differences should never be nested in application code,
and it should be easy to tell when there may be differences - by recognizing
that config is being imported.